Update api/global benchmarks
Move SDK related tests to SDK where applicable
